reworking bdd interface
